updated vgg_contrib.

+++ flexbox/src/flexbox_node.cpp
@@ -133,7 +133,7 @@ bool flexbox_node::add_child(uint32_t id)
 
 bool flexbox_node::add_child(p_node &child, uint32_t id)
 {
-    auto count = YGNodeGetChildCount(this->node_);
+    auto count = static_cast<uint32_t>(YGNodeGetChildCount(this->node_));
     if (id == static_cast<uint32_t>(-1))
     {
         id = count;
